:recycle: s3-integration extract rest
