+++ refarch-integrations/refarch-s3-integration/refarch-s3-integration-core/src/main/java/de/muenchen/refarch/integration/s3/domain/validation/FolderInFilePathValidator.java
@@ -1,4 +1,4 @@
-package de.muenchen.refarch.integration.s3.adapter.in.rest.validation;
+package de.muenchen.refarch.integration.s3.domain.validation;
 
 import jakarta.validation.ConstraintValidator;
 import jakarta.validation.ConstraintValidatorContext;
