+++ refarch-integrations/refarch-s3-integration/refarch-s3-integration-rest-core/src/main/java/de/muenchen/refarch/integration/s3/adapter/in/rest/mapper/PresignedUrlMapper.java
