+++ refarch-integrations/refarch-s3-integration/refarch-s3-integration-rest-starter/src/main/java/de/muenchen/refarch/integration/s3/configuration/S3IntegrationRestAutoConfiguration.java
@@ -1,0 +1,9 @@
+package de.muenchen.refarch.integration.s3.configuration;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+@RequiredArgsConstructor
+public class S3IntegrationRestAutoConfiguration {
+}
