+++ refarch-integrations/refarch-s3-integration/refarch-s3-integration-rest-starter/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,0 +1,1 @@
+de.muenchen.refarch.integration.s3.configuration.S3IntegrationRestAutoConfiguration
